Fix DocumentDb + testing DocumentDb

+++ implementation/src/main/java/database/DB_DocumentManager.java
@@ -88,6 +88,9 @@ public class DB_DocumentManager extends DB_Controller implements DB_DocumentMana
      */
     @Override
     public boolean deleteDocument(String name) {
+        if (!this.isNameAlreadyUsed(name)) {
+            return false;
+        }
         this.openConnection();
         String sqlstatement = "DELETE FROM documents WHERE documentName = ?";
         try (PreparedStatement stmt = connection.prepareStatement(sqlstatement)) {
@@ -112,6 +115,9 @@ public class DB_DocumentManager extends DB_Controller implements DB_DocumentMana
      */
     @Override
     public boolean updateDocument(String oldName, String newName) {
+        if (!this.isNameAlreadyUsed(oldName)) {
+            return false;
+        }
         this.openConnection();
         String revisionNumber = "SELECT revision FROM documents WHERE documentName = ?";
         String sqlstatement = "UPDATE documents SET revision = ? , "
